Add success check to http requests
Added a check so that the finished callbacks provided in dohttpreq are only
called if curl did not encounter errors during the request. This helps avoiding
issues trying to parse invalid data and is generally the better way to handle it
as you would expect the download finished callback to only be called on success.
Another change is to cause http errors to be considered as a request fail which
also makes sense considering you probably only care about valid data.