Fix database hooks with absolute IDs doing nothing
Previously when a database hook (in Wren) was made with an absolute ID -
that is a '@' followed by a 64-bit hex hash value - it would be parsed
as a signed 64-bit value and if it exceeded 0x7fffffff then it would be
capped at that, giving the hook an invalid ID and stopping the desired
asset from hitting it.

This also meant that if you had multiple assets with names or types that
fell into this, they'd be treated as the same and you'd get a collision
error between them.