Fix the game crashing if the basemod is not installed
When SBLT tried to load the Wren runtime, it would check if the
wren entry point file exists. If not it wouldn't load the runtime, as
there's nothing it would do.

This was not checked when applying a tweak, leading to a crash on startup
if the basemod wasn't present.

Fixes #68.