When loading assets, handle end-of-bundle assets - Fixes #73
When we read the DB headers we're not told the length of each asset -
rather we get their offsets and use that to calculate the length.

This doesn't work for the last asset, and previously the game would
crash when attempting to read one. This patch fixes that, calculating
the file length at read time.

This is done for all three places the length is accessed - reading the
file from Lua and Wren and when setting up a redirected in-bundle asset.