Fix crashes on computers with all.blb
Previously, when we went through all the files in the assets directory
we checked if they ended with _h.bundle - to do this we used std::string::compare
to check the last nine characters of the filename.

This threw an exception if the file was less than nine characters long, and old
versions of PD2 seem to have had an all.blb file - this being too short would
trigger the exception, which then went unhandled.

This patch now checks the filename length to avoid this crash, skipping
files that are too short since they're obviously not bundle headers.